Move visitors to valuetype directory
