+++ libs/execution/src/lib/index.ts
@@ -6,7 +6,7 @@ export * from './blocks';
 export * from './constraints';
 export * from './types';
 export * from './util';
-export * from './visitors';
+export * from './types/valuetypes/visitors';
 
 export * from './execution-context';
 export * from './extension';
